pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ts-foundation-benchmark"
version = "1.0.0"
description = "Benchmark comparing classical forecasting against time-series foundation models"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"

dependencies = [
    "numpy>=1.24.0",
    "statsmodels>=0.14.0",
    "xgboost>=2.0.0",
    "torch>=2.0.0",
]

[project.optional-dependencies]
chronos = ["chronos-forecasting>=1.0.0"]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
disallow_untyped_defs = true
